requirements.txt
boto3==1.4.7
botocore==1.7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
daphne==1.3.0
Django==1.10.5
django-cors-headers==2.0.2
django-extensions==1.9.6
django-filter==1.0.4
django-filters==0.2.1
django-jenkins==0.110.0
django-redis==4.8.0
django-rest-swagger==2.1.2
django-storages==1.6.5
djangorestframework==3.5.4
docutils==0.14
flake8==3.4.1
gunicorn==19.7.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==17.3.1
incremental==17.5.0
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
lazy-object-proxy==1.3.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
openapi-codec==1.3.2
packaging==16.8
pep8==1.7.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pyparsing==2.1.10
pytz==2017.3
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
simplejson==3.11.1
six==1.10.0
Twisted==17.9.0
txaio==2.8.2
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
zope.interface==4.4.3
